http.c share public function for both client and server http protocol
